switch from nose to pytest (and pytest-cov), see also #70
pytest is a more advanced testing framework than nose (IMHO) and makes writing tests a
better experience as one has to type less and has more powerful mechanisms at hand.

Also, in case of a test failure, the output is more helpful in finding the cause.

See: http://pytest.org/latest/

With this changeset, the amount of tests and test fails (none for normal tests, some complaints for pylint) has not changed.

Currently, pytest runs in "compatibility mode", so it is happily executing your existing "written-for-nose" tests.

But future tests (and also existing tests) can now be written with less and prettier code.