talk to servers on non-standard https port (like -s acme.example.org:8443)
note:

I removed Network.server attribute, it was not used (only .server_url is used).

I also removed the sanity-check for the ACME server (so one can use server:port syntax, which was considered invalid).
I don't think this sanity-check is needed, because in the end, we can't fully check if it is correct anyway (you can always
give valid, but non-working server names or ports).
Also, many people will use the default ACME server, which is builtin and correct.