crypto_util: fix bug in crypto_util.pyopenssl_x509_name_as_text()
crypto_util.pyopenssl_x509_name_as_text() returns the "/".join() of strings of the form "key=value".
It gets these as an array of tuples.

The bug:
During iteration, a string "{0}={1}" was supposed to be formatted with the key and value.
Unfortunately, the format string was never actually formatted with `key` or `value`.

The fix:
Call `.format()` on the format string in the iteration.