Do not reinstall what's already installed
There's no need to reinstall all packages. Few of  them are quite heavy (gcc) and most  of the time, when it comes to developers, already available. You need git, for instance, to clone the repository itself, that's going to be there and configured already too.

Please consider this change so that bootstrapping is going to be even faster. Best Regards