le-auto: clarify logic in DeterminePythonVersion.
In code that follows this statements, all exports were just expansions
of LE_PYTHON with a undefined-or-empty fallback to the found ones.

This is quite weird given that we are more willing to use the found
one. In respect to the original behavior, I also made the original
LE_PYTHON value our first candidate to try.