Add MacPorts support to Mac bootstrap
Why? MacPorts is a popular and venerable package manager for OS X.
LetsEncrypt should use it if it is already installed.

This change modifies the Mac bootstrap script to add support for
MacPorts package manager. HomeBrew support remains, and if neither
package manager is installed, the bootstrap script will still
default to installing HomeBrew.