Increase Security & Reduce Payload When Using Git
`--depth=1 --branch=master` The depth option will make sure to copy the least bit of history possible. This reduces the payload on the serving host (GitHub) and not requesting the entire git history to minimize the footprint on the implemented server.

Adding ` letsencrypt` to the end of the clone ensures the repo is cloned into its own directory and is accessible by the next command:

`rm -rf !$/.git` Removes the '.git' directory from the 'letsencrypt' directory making it not a Git repository any more. This, more importantly, improves security for both the user and the repo. Without getting into details (_that are beyond me_); Studies have shown that most breaches into servers and repo's are caused by leaving the .git commit history on production servers that "can contain" sensitive data. To be safe, lets just delete it.