Simple Installer
Allows User's to download and install using `wget` to safeguard User's production
servers from unknown vulnerabilities and reduce disc footprint by eliminating the
need to `git clone` the project.

Why is this change necessary?

Using git as an installer opens risks of sensitive data discovery within the
'.git' directory and development tools can contain zero-day vulnerabilities,
especially (unknowingly) with testing scripts. Also, the usage of downloading
with git and installing from the local repo is heavy and bloated. Currently
this occupies a 12.3M footprint on the User's disc and 7.8M of this is in
the `.git` directory.

> Security best practices says, "Leave Git only for the Developers and not the End-Users".

How does this request address the issue?

 1. Create simple bash.sh that can be downloaded and executed using `wget | bash`
 2. List non-essential files to the app with `export-ignore` in '.gitattributes'

Changes to be merged:

 * modified: .gitattributes
 * modified: README.rst
 * new file: tools/installer

What side effects does this change have?

 * Each release, the `<RELEASE_VERSION>` field will need updating ('installer')

Conclusion

Using this mothod of delivering the `letsencrypt` app will make it safer for both
the End-User and Let's Encrypt by eliminating un-needed risk and liability. The
total size of the app is greatly reduced from 12.3M down to 2.9M on the server.
As a suggestion, the use of a URL shortener service (goo.gl for an example) will
make the `wget` script less frightening and you guys can track who/what/where of
all the downloads!

See: http://git-scm.com/book/en/v2/Customizing-Git-Git-Attributes#Exporting-Your-Repository