Official npm Installer
Allows User's to download and install using an official npm package.

Why is this Proposal Necessary?

From what I can find there is no official 'letsencrypt' package on npm. There are a
few variations by the works of coolaj86. However, they are written in js and "over-built"
by requiring several `dependencies` equating to over a hundred sub-dependencies. This is
not nessessary and node can utilized the current scrpts in this repository.

How does this request address the issue?

 1. Create an node package.json in the root
 2. Add a '.npmignore' to eliminate bloat
 3. Add a simple 'npm-cli' binary that executes './letsencrypt' and passes the args to it
 4. Create an 'official' "letsencrypt" organization and package(s) on npm

Changes to be committed:

 * new file:   .npmignore
 * modified:   README.rst
 * new file:   bin/npm-cli
 * new file:   package.json

Usage:

 * Installation:

   user@webserver:~$ npm i -g @letsencrypt/letsencrypt

 > The normally used './letsencrypt-auto <arg(s)>' command can be now executed globaly
 > using `letsencrypt <arg(s)>' in a node environment.

 > **_NOTE:_** A user will have to manually symlink the downloaded certs (See TODO).

What side effects does this change have?

 * Repo maintainers will have to update the 'version' key within the 'package.json'
   file with each release (_there are tool that can automate this_)
 * A npm organization will need to be created/maintained by this repo's owners

Conclusion

This will allow node users to easily install the "Let's Encrypt" software via a npm
package that is officially maintained by it's creators and insuring that there are no
dependencies that can become outdated or have vulnerabilities.

TODO: [npm-cli] Add function to take an arg to istruct where to symlink the certs.