Auto creates a vhost configuration file if possible
When the target vhost is not configured in a separate file, the config is
written to the root config file (nginx.conf).

This change replaces this behavior by auto-detecting the vhost configuration
directory based on the existing vhost configuration files. This way the vhost
configuration files is created in a separate directory.