Don't error if the current_target is a hard file
In at least some scenarios when you run

```
letsencrypt certonly --webroot -w <directory> -d <domain>
```

The client exits with the error "OSError: [Errno 22] Invalid argument:
'/etc/letsencrypt/live/<other-domain>/cert.pem'". Particularly confusing is the
fact that this error message is from letsencrypt 1) attempting to read a file
that exists and 2) attempting to read a domain that's not related to the domain
being validated.

The error comes from the letsencrypt client attempting to find duplicate
certs, and resolve symlinks by calling `os.readlink`. However, in some cases
letsencrypt attempts to call `os.readlink` on a hard file, which throws an
error in Python.

Since `current_target` is only attempting to retrieve the path to the given
certificate, it seems safe enough to check if a file is a symlink before
attempting to dereference it with `os.readlink`.

I verified that this patch fixes the issue on my own server - without the
patch, I get the error in #2054 and with the patch, I'm able to get a
certificate.

Fixes https://github.com/letsencrypt/letsencrypt/issues/2054.