Fixes Certbot for OS X
This commit removes the condition that checks for Python's path. The hard-coded path is incorrect since in OS X 10.11 Python is installed in /usr/bin/python. Removing the condition makes sense because any of the software we use to install Python should return a "Already installed" message if Python is installed