add -D/--webroot-delay-auth option
With this new option enabled, the "webroot" plugin will wait for user
input before the challanges will be verified.

This is useful for creating certificates for shared hosting plans
without root (or even no SSH access at all): The challenges can be
proposed from a foreign machine, and with this option, the user has
the time to copy the validation files to the actual webserver before
the challanges are verified.