Add clean up function to integration test script.
Refine check for abandoned web server process.

There are two possible causes for web servers to be running, which require
two slightly different responses. When certbot is run in manual mode it used
to leave a web server running in an abandoned process, now there is a function
to check for that and kill it if it is running. Secondly, if either of the
first two calls to certbot in the integration test fail, then the script will
exit before killing the web servers that the integration test spins up. There
is a separate function to check for that and clean it up.