Add a trap to cleanup local server used in testing
The `boulder-integration.sh` script starts a test server for each of the
first two calls to `common`. If either of these calls fail (i.e. because
certbot can't connect to a local boulder instance because of incorrect firewall
settings or because a local boulder instance isn't running) then the
server is left running. This will cause subsequent runs of
`boulder-integration.sh` to fail with "[Errno 98] Address already in
use".