During certbot-auto upgrade: update venv if it exists
When `certbot-auto` detects it should upgrade, it used to just delete the whole venv folder, and re-create it.
This caused issues with manually installed plugins that just disappeared.

With this change, instead of deleting the whole venv folder, we keep it as-is, and let virtualenv to do its thing.

Fixes #4037