Don't install argparse on python >=3.2
On Python 3.5, I got an exception about argparse.ArgumentParser
not supporting allow_abbrev in __init__.
Turns out argparse is in stdlib on Python >=3.2, so no need to install
argparse there.
Worse, installing argparse from PyPi pulls in a stale version that does
not support the argument allow_abbrev.
ConfigArgParse.ArgumentParser, which uses as argparse.ArgumentParser
as base class, assumes 'allow_abbrev' is supported since we are on
Python 3.5(where this actual argument got supported).
And with the old class installed we get the exception.