Update example CloudFlare authenticator script
This change removes the fairly broken regex for getting the zone name from the domain/subdomain. 

The example will now work with root domains such as `example.com` (previously `DOMAIN` was a blank string).

For CNAME-d subdomains where `sub.example1.com` -> `another-sub.example2.com`, a possible caveat is that`DOMAIN` will be `example2.com` not `example1.com`. It appears that `certbot` checks both domains anyway (I think?) but I believe returning the SOA domain is preferable because it's likely to be the origin host generating this type of certificate, not the person who set up the CNAME (if they are different people).