Avoid crashing on SIGWINCH
When terminals are resized, the terminal will usually send a SIGWINCH -
the Window Change signal. This usually gives interactive applications a
chance to render the UI with the new width and height. However, if this
sinal isn't handled or trapped, the default behavior tends to be
aborting the process.

This change hooks SIGWINCH up to SIG_IGN, which ignores the signal. This
is safe, because we don't really care that the terminal has been
resized.