Let certbot-auto work behind the Great Firewall of China. Fix #3222. Ref #4371.
...without needing to get pip >= 8 by hand first and without needing to muck with pip.conf.

We do it by upgrading to a new version of pipstrap which understands pip's own PIP_INDEX_URL env var for specifying a mirror. That done, certbot-auto simply sets that var to a known Chinese mirror in response to a CLI flag, and off we go.

This could be made automatic in the future: see 2nd-last paragraph of https://github.com/erikrose/pipstrap/pull/10#issuecomment-333637736. But there's no reason not to land this simple version first.