Add back macOS and Boulder for test-everything.
This commit should not be merged to master (if that happens, just revert it).

In https://github.com/certbot/certbot/pull/5270, we removed the macOS and
most Boulder integration tests from .travis.yml because they made builds very
slow. We still want to run those tests nightly, though. The test-everything
branch will have the tests that are too expensive to run for every PR.