Support overriding challenge domain in DNS plugins
Support a new option --dns-<plugin>-override-challenge to
instruct certbot to set up the TXT record for a DNS-01 challenge
at an alternate location instead of the default _acme-challenge.<domain>.

Independently of this option, this change also incidentally fixes
the possibly uncommon use case where _acme-challenge is itself a
delegation apex (previously the RFC2136 plugin would generate an
update for the wrong zone in that case).

Refactor common code across test modules for the various DNS plugins
into a new method in dns_test_common.BaseAuthenticatorTest.

As a consequence, bump the oldest requirement of all DNS plugins to
the latest certbot core.

As a consequence, bump their oldest requirement for acme to 0.25.0,
as current certbot requires acme.magic_typing.