Fix Apache configurator test
One of the Apache configurator tests was relying on the order in which
calls are made by ensure_listen() to parser.add_dir(). Under the hood,
the order is determined by creating a list and a set, adding some items
to the set, and then iterating over the difference (itself a set) between
the set and the list. Unfortunately, sets are unordered collections, and
iteration order is therefore not guaranteed. This resulted in test
failures when the iteration order was not as the test writer had
anticipated. We deal with this by instead building a list of all the
calls to parser_dir() and comparing it to a list of expected calls.
This has the extra benefit of ensuring that no extraneous calls are
added and no expected calls are missing.